Allow setting service-specific account settings via Account
Previously, service-specific account configuration values could only
be set via the ServiceAccount interface.  This commit adds API so that
the Account interface can now read and set service-specific settings
also.

In the latest libaccounts-qt (v1.4) better support for GVariant to
QVariant conversion was added, and so my hand rolled conversion
function is no longer required.  This commit removes that in favour
of using the improved API offered by libaccounts-qt in 1.4.

When setting service-specific settings, the enabled status of the
account with that service gets overwritten.  Thus, we need to enable
or disable the service after setting service-specific settings.

Finally, this commit also improves various unit tests related to
service-specific account settings and enabledServiceNames.